cleanup temporal implementation
