+++ internal/temporal/workflows/calculation_workflow.go
@@ -4,44 +4,21 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/flexprice/flexprice/internal/temporal/models"
 	"go.temporal.io/sdk/workflow"
 )
 
-type CalculationResult struct {
-	InvoiceID   string
-	TotalAmount float64
-	Items       []InvoiceItem
-}
-
-type InvoiceItem struct {
-	Description string
-	Amount      float64
-}
-
-// Activity result types
-type FetchDataActivityResult struct {
-	CustomerData interface{}
-	PlanData     interface{}
-	UsageData    interface{}
-}
-
-type CalculateActivityResult struct {
-	InvoiceID   string
-	TotalAmount float64
-	Items       []InvoiceItem
-}
-
-func CalculateChargesWorkflow(ctx workflow.Context, input BillingWorkflowInput) (*CalculationResult, error) {
+func CalculateChargesWorkflow(ctx workflow.Context, input models.BillingWorkflowInput) (*models.CalculationResult, error) {
 	logger := workflow.GetLogger(ctx)
 	logger.Info("Starting calculation workflow",
 		"customerID", input.CustomerID,
 		"subscriptionID", input.SubscriptionID)
 
 	// Just return a mock result for now
-	return &CalculationResult{
+	return &models.CalculationResult{
 		InvoiceID:   fmt.Sprintf("INV-%s-%d", input.CustomerID, time.Now().Unix()),
 		TotalAmount: 100.0,
-		Items: []InvoiceItem{
+		Items: []models.InvoiceItem{
 			{
 				Description: "Mock charge",
 				Amount:      100.0,
